turbojpeg_sys

Function tj3Compress16

pub unsafe extern "C" fn tj3Compress16(
    handle: tjhandle,
    srcBuf: *const c_ushort,
    width: c_int,
    pitch: c_int,
    height: c_int,
    pixelFormat: c_int,
    jpegBuf: *mut *mut c_uchar,
    jpegSize: *mut size_t,
) -> c_int

Compress a packed-pixel RGB, grayscale, or CMYK image with 13 to 16 bits of data precision per sample into a lossless JPEG image with the same data precision.

@param handle handle to a TurboJPEG instance that has been initialized for compression

@param srcBuf pointer to a buffer containing a packed-pixel RGB, grayscale, or CMYK source image to be compressed. This buffer should normally be pitch * height samples in size. However, you can also use this parameter to compress from a specific region of a larger buffer. The data precision of the source image (from 13 to 16 bits per sample) can be specified using #TJPARAM_PRECISION and defaults to 16 if #TJPARAM_PRECISION is unset or out of range.

@param width width (in pixels) of the source image

@param pitch samples per row in the source image. Normally this should be width * #tjPixelSize[pixelFormat], if the image is unpadded. (Setting this parameter to 0 is the equivalent of setting it to width * #tjPixelSize[pixelFormat].) However, you can also use this parameter to specify the row alignment/padding of the source image, to skip rows, or to compress from a specific region of a larger buffer.

@param height height (in pixels) of the source image

@param pixelFormat pixel format of the source image (see @ref TJPF “Pixel formats”.)

@param jpegBuf address of a pointer to a byte buffer that will receive the JPEG image. TurboJPEG has the ability to reallocate the JPEG buffer to accommodate the size of the JPEG image. Thus, you can choose to: -# pre-allocate the JPEG buffer with an arbitrary size using #tj3Alloc() and let TurboJPEG grow the buffer as needed, -# set *jpegBuf to NULL to tell TurboJPEG to allocate the buffer for you, or -# pre-allocate the buffer to a “worst case” size determined by calling #tj3JPEGBufSize() and adding the return value to the size of the ICC profile (if any) that was previously associated with the TurboJPEG instance (see #tj3SetICCProfile().) This should ensure that the buffer never has to be re-allocated. (Setting #TJPARAM_NOREALLOC guarantees that it won’t be.) . If you choose option 1 or 3, then *jpegSize should be set to the size of your pre-allocated buffer. In any case, unless you have set #TJPARAM_NOREALLOC, you should always check *jpegBuf upon return from this function, as it may have changed.

@param jpegSize pointer to a size_t variable that holds the size of the JPEG buffer. If *jpegBuf points to a pre-allocated buffer, then *jpegSize should be set to the size of the buffer. Upon return, *jpegSize will contain the size of the JPEG image (in bytes.) If *jpegBuf points to a JPEG buffer that is being reused from a previous call to one of the JPEG compression functions, then *jpegSize is ignored.

@return 0 if successful, or -1 if an error occurred (see #tj3GetErrorStr() and #tj3GetErrorCode().)