turbojpeg_sys

Function tj3DecodeYUV8

pub unsafe extern "C" fn tj3DecodeYUV8(
    handle: tjhandle,
    srcBuf: *const c_uchar,
    align: c_int,
    dstBuf: *mut c_uchar,
    width: c_int,
    pitch: c_int,
    height: c_int,
    pixelFormat: c_int,
) -> c_int

Decode an 8-bit-per-sample unified planar YUV image into an 8-bit-per-sample packed-pixel RGB or grayscale image. This function performs color conversion (which is accelerated in the libjpeg-turbo implementation) but does not execute any of the other steps in the JPEG decompression process.

@param handle handle to a TurboJPEG instance that has been initialized for decompression

@param srcBuf pointer to a buffer containing a unified planar YUV source image to be decoded. The size of this buffer should match the value returned by #tj3YUVBufSize() for the given image width, height, row alignment, and level of chrominance subsampling (see #TJPARAM_SUBSAMP.) The Y, U (Cb), and V (Cr) image planes should be stored sequentially in the source buffer. (Refer to @ref YUVnotes “YUV Image Format Notes”.)

@param align row alignment (in bytes) of the YUV source image (must be a power of 2.) Setting this parameter to n indicates that each row in each plane of the YUV source image is padded to the nearest multiple of n bytes (1 = unpadded.)

@param dstBuf pointer to a buffer that will receive the packed-pixel decoded image. This buffer should normally be pitch * height bytes in size. However, you can also use this parameter to decode into a specific region of a larger buffer.

@param width width (in pixels) of the source and destination images

@param pitch bytes per row in the destination image. Normally this should be set to width * #tjPixelSize[pixelFormat], if the destination image should be unpadded. (Setting this parameter to 0 is the equivalent of setting it to width * #tjPixelSize[pixelFormat].) However, you can also use this parameter to specify the row alignment/padding of the destination image, to skip rows, or to decode into a specific region of a larger buffer.

@param height height (in pixels) of the source and destination images

@param pixelFormat pixel format of the destination image (see @ref TJPF “Pixel formats”.)

@return 0 if successful, or -1 if an error occurred (see #tj3GetErrorStr() and #tj3GetErrorCode().)