turbojpeg_sys

Function tj3Decompress8

pub unsafe extern "C" fn tj3Decompress8(
    handle: tjhandle,
    jpegBuf: *const c_uchar,
    jpegSize: size_t,
    dstBuf: *mut c_uchar,
    pitch: c_int,
    pixelFormat: c_int,
) -> c_int

Decompress a JPEG image with 2 to 8 bits of data precision per sample into a packed-pixel RGB, grayscale, or CMYK image with the same data precision. The @ref TJPARAM “parameters” that describe the JPEG image will be set when this function returns.

@param handle handle to a TurboJPEG instance that has been initialized for decompression

@param jpegBuf pointer to a byte buffer containing the JPEG image to decompress

@param jpegSize size of the JPEG image (in bytes)

@param dstBuf pointer to a buffer that will receive the packed-pixel decompressed image. This buffer should normally be pitch * destinationHeight samples in size. However, you can also use this parameter to decompress into a specific region of a larger buffer. NOTE: If the JPEG image is lossy, then destinationHeight is either the scaled JPEG height (see #TJSCALED(), #TJPARAM_JPEGHEIGHT, and #tj3SetScalingFactor()) or the height of the cropping region (see #tj3SetCroppingRegion().) If the JPEG image is lossless, then destinationHeight is the JPEG height.

@param pitch samples per row in the destination image. Normally this should be set to destinationWidth * #tjPixelSize[pixelFormat], if the destination image should be unpadded. (Setting this parameter to 0 is the equivalent of setting it to destinationWidth * #tjPixelSize[pixelFormat].) However, you can also use this parameter to specify the row alignment/padding of the destination image, to skip rows, or to decompress into a specific region of a larger buffer. NOTE: If the JPEG image is lossy, then destinationWidth is either the scaled JPEG width (see #TJSCALED(), #TJPARAM_JPEGWIDTH, and #tj3SetScalingFactor()) or the width of the cropping region (see #tj3SetCroppingRegion().) If the JPEG image is lossless, then destinationWidth is the JPEG width.

@param pixelFormat pixel format of the destination image (see @ref TJPF “Pixel formats”.)

@return 0 if successful, or -1 if an error occurred (see #tj3GetErrorStr() and #tj3GetErrorCode().)