turbojpeg_sys

Function tj3DecompressHeader

pub unsafe extern "C" fn tj3DecompressHeader(
    handle: tjhandle,
    jpegBuf: *const c_uchar,
    jpegSize: size_t,
) -> c_int

Retrieve information about a JPEG image without decompressing it, or prime the decompressor with quantization and Huffman tables. If a JPEG image is passed to this function, then the @ref TJPARAM “parameters” that describe the JPEG image will be set when the function returns. If a JPEG image is passed to this function and #TJPARAM_SAVEMARKERS is set to 2 or 4, then the ICC profile (if any) will be extracted from the JPEG image. (#tj3GetICCProfile() can then be used to retrieve the profile.)

@param handle handle to a TurboJPEG instance that has been initialized for decompression

@param jpegBuf pointer to a byte buffer containing a JPEG image or an “abbreviated table specification” (AKA “tables-only”) datastream. Passing a tables-only datastream to this function primes the decompressor with quantization and Huffman tables that can be used when decompressing subsequent “abbreviated image” datastreams. This is useful, for instance, when decompressing video streams in which all frames share the same quantization and Huffman tables.

@param jpegSize size of the JPEG image or tables-only datastream (in bytes)

@return 0 if successful, or -1 if an error occurred (see #tj3GetErrorStr() and #tj3GetErrorCode().)