turbojpeg_sys

Function tj3DecompressToYUVPlanes8

pub unsafe extern "C" fn tj3DecompressToYUVPlanes8(
    handle: tjhandle,
    jpegBuf: *const c_uchar,
    jpegSize: size_t,
    dstPlanes: *mut *mut c_uchar,
    strides: *mut c_int,
) -> c_int

Decompress an 8-bit-per-sample JPEG image into separate 8-bit-per-sample Y, U (Cb), and V (Cr) image planes. This function performs JPEG decompression but leaves out the color conversion step, so a planar YUV image is generated instead of a packed-pixel image. The @ref TJPARAM “parameters” that describe the JPEG image will be set when this function returns.

@param handle handle to a TurboJPEG instance that has been initialized for decompression

@param jpegBuf pointer to a byte buffer containing the JPEG image to decompress

@param jpegSize size of the JPEG image (in bytes)

@param dstPlanes an array of pointers to Y, U (Cb), and V (Cr) image planes (or just a Y plane, if decompressing a grayscale image) that will receive the decompressed image. These planes can be contiguous or non-contiguous in memory. Use #tj3YUVPlaneSize() to determine the appropriate size for each plane based on the scaled JPEG width and height (see #TJSCALED(), #TJPARAM_JPEGWIDTH, #TJPARAM_JPEGHEIGHT, and #tj3SetScalingFactor()), strides, and level of chrominance subsampling (see #TJPARAM_SUBSAMP.) Refer to @ref YUVnotes “YUV Image Format Notes” for more details.

@param strides an array of integers, each specifying the number of bytes per row in the corresponding plane of the YUV image. Setting the stride for any plane to 0 is the same as setting it to the scaled plane width (see @ref YUVnotes “YUV Image Format Notes”.) If strides is NULL, then the strides for all planes will be set to their respective scaled plane widths. You can adjust the strides in order to add an arbitrary amount of row padding to each plane or to decompress the JPEG image into a subregion of a larger planar YUV image.

@return 0 if successful, or -1 if an error occurred (see #tj3GetErrorStr() and #tj3GetErrorCode().)