turbojpeg_sys

Function tj3EncodeYUVPlanes8

pub unsafe extern "C" fn tj3EncodeYUVPlanes8(
    handle: tjhandle,
    srcBuf: *const c_uchar,
    width: c_int,
    pitch: c_int,
    height: c_int,
    pixelFormat: c_int,
    dstPlanes: *mut *mut c_uchar,
    strides: *mut c_int,
) -> c_int

Encode an 8-bit-per-sample packed-pixel RGB or grayscale image into separate 8-bit-per-sample Y, U (Cb), and V (Cr) image planes. This function performs color conversion (which is accelerated in the libjpeg-turbo implementation) but does not execute any of the other steps in the JPEG compression process.

@param handle handle to a TurboJPEG instance that has been initialized for compression

@param srcBuf pointer to a buffer containing a packed-pixel RGB or grayscale source image to be encoded. This buffer should normally be pitch * height bytes in size. However, you can also use this parameter to encode from a specific region of a larger buffer.

@param width width (in pixels) of the source image

@param pitch bytes per row in the source image. Normally this should be width * #tjPixelSize[pixelFormat], if the image is unpadded. (Setting this parameter to 0 is the equivalent of setting it to width * #tjPixelSize[pixelFormat].) However, you can also use this parameter to specify the row alignment/padding of the source image, to skip rows, or to encode from a specific region of a larger packed-pixel image.

@param height height (in pixels) of the source image

@param pixelFormat pixel format of the source image (see @ref TJPF “Pixel formats”.)

@param dstPlanes an array of pointers to Y, U (Cb), and V (Cr) image planes (or just a Y plane, if generating a grayscale image) that will receive the encoded image. These planes can be contiguous or non-contiguous in memory. Use #tj3YUVPlaneSize() to determine the appropriate size for each plane based on the image width, height, strides, and level of chrominance subsampling (see #TJPARAM_SUBSAMP.) Refer to @ref YUVnotes “YUV Image Format Notes” for more details.

@param strides an array of integers, each specifying the number of bytes per row in the corresponding plane of the YUV image. Setting the stride for any plane to 0 is the same as setting it to the plane width (see @ref YUVnotes “YUV Image Format Notes”.) If strides is NULL, then the strides for all planes will be set to their respective plane widths. You can adjust the strides in order to add an arbitrary amount of row padding to each plane or to encode an RGB or grayscale image into a subregion of a larger planar YUV image.

@return 0 if successful, or -1 if an error occurred (see #tj3GetErrorStr() and #tj3GetErrorCode().)