turbojpeg_sys

Function tj3LoadImage8

pub unsafe extern "C" fn tj3LoadImage8(
    handle: tjhandle,
    filename: *const c_char,
    width: *mut c_int,
    align: c_int,
    height: *mut c_int,
    pixelFormat: *mut c_int,
) -> *mut c_uchar

Load a packed-pixel image with 2 to 8 bits of data precision per sample from disk into memory.

@param handle handle to a TurboJPEG instance

@param filename name of a file containing a packed-pixel image in Windows BMP or PBMPLUS (PPM/PGM) format. Windows BMP files require 8-bit-per-sample data precision. When loading a PBMPLUS file, the target data precision (from 2 to 8 bits per sample) can be specified using #TJPARAM_PRECISION and defaults to 8 if #TJPARAM_PRECISION is unset or out of range. If the data precision of the PBMPLUS file does not match the target data precision, then upconverting or downconverting will be performed.

@param width pointer to an integer variable that will receive the width (in pixels) of the packed-pixel image

@param align row alignment (in samples) of the packed-pixel buffer to be returned (must be a power of 2.) Setting this parameter to n will cause all rows in the buffer to be padded to the nearest multiple of n samples (1 = unpadded.)

@param height pointer to an integer variable that will receive the height (in pixels) of the packed-pixel image

@param pixelFormat pointer to an integer variable that specifies or will receive the pixel format of the packed-pixel buffer. The behavior of this function varies depending on the value of *pixelFormat passed to the function:

@ref TJPF_UNKNOWN : The packed-pixel buffer returned by this function will use the most optimal pixel format for the file type, and *pixelFormat will contain the ID of that pixel format upon successful return from this function.

@ref TJPF_GRAY : Only PGM files and 8-bit-per-pixel BMP files with a grayscale colormap can be loaded.

@ref TJPF_CMYK : The RGB or grayscale pixels stored in the file will be converted using a quick & dirty algorithm that is suitable only for testing purposes. (Proper conversion between CMYK and other formats requires a color management system.)

Other @ref TJPF “pixel formats” : The packed-pixel buffer will use the specified pixel format, and pixel format conversion will be performed if necessary.

@return a pointer to a newly-allocated buffer containing the packed-pixel image, converted to the chosen pixel format and with the chosen row alignment, or NULL if an error occurred (see #tj3GetErrorStr().) This buffer should be freed using #tj3Free().