turbojpeg_sys

Function tj3SaveImage16

pub unsafe extern "C" fn tj3SaveImage16(
    handle: tjhandle,
    filename: *const c_char,
    buffer: *const c_ushort,
    width: c_int,
    pitch: c_int,
    height: c_int,
    pixelFormat: c_int,
) -> c_int

Save a packed-pixel image with 13 to 16 bits of data precision per sample from memory to disk.

@param handle handle to a TurboJPEG instance

@param filename name of a file to which to save the packed-pixel image, which will be stored in PBMPLUS (PPM/PGM) format. The source data precision (from 13 to 16 bits per sample) can be specified using #TJPARAM_PRECISION and defaults to 16 if #TJPARAM_PRECISION is unset or out of range.

@param buffer pointer to a buffer containing a packed-pixel RGB, grayscale, or CMYK image to be saved

@param width width (in pixels) of the packed-pixel image

@param pitch samples per row in the packed-pixel image. Setting this parameter to 0 is the equivalent of setting it to width * #tjPixelSize[pixelFormat].

@param height height (in pixels) of the packed-pixel image

@param pixelFormat pixel format of the packed-pixel image (see @ref TJPF “Pixel formats”.) If this parameter is set to @ref TJPF_GRAY, then the image will be stored in PGM format. Otherwise, the image will be stored in PPM format. If this parameter is set to @ref TJPF_CMYK, then the CMYK pixels will be converted to RGB using a quick & dirty algorithm that is suitable only for testing purposes. (Proper conversion between CMYK and other formats requires a color management system.)

@return 0 if successful, or -1 if an error occurred (see #tj3GetErrorStr().)