turbojpeg_sys

Function tj3SetICCProfile

pub unsafe extern "C" fn tj3SetICCProfile(
    handle: tjhandle,
    iccBuf: *mut c_uchar,
    iccSize: size_t,
) -> c_int

Embed an ICC (International Color Consortium) color management profile in JPEG images generated by subsequent compression and lossless transformation operations.

@param handle handle to a TurboJPEG instance that has been initialized for compression

@param iccBuf pointer to a byte buffer containing an ICC profile. A copy is made of the ICC profile, so this buffer can be freed or reused as soon as this function returns. Setting this parameter to NULL or setting iccSize to 0 removes any ICC profile that was previously associated with the TurboJPEG instance.

@param iccSize size of the ICC profile (in bytes.) Setting this parameter to 0 or setting iccBuf to NULL removes any ICC profile that was previously associated with the TurboJPEG instance.

@return 0 if successful, or -1 if an error occurred (see #tj3GetErrorStr().)