turbojpeg_sys

Function tj3Transform

pub unsafe extern "C" fn tj3Transform(
    handle: tjhandle,
    jpegBuf: *const c_uchar,
    jpegSize: size_t,
    n: c_int,
    dstBufs: *mut *mut c_uchar,
    dstSizes: *mut size_t,
    transforms: *const tjtransform,
) -> c_int

Losslessly transform a JPEG image into another JPEG image. Lossless transforms work by moving the raw DCT coefficients from one JPEG image structure to another without altering the values of the coefficients. While this is typically faster than decompressing the image, transforming it, and re-compressing it, lossless transforms are not free. Each lossless transform requires reading and performing entropy decoding on all of the coefficients in the source image, regardless of the size of the destination image. Thus, this function provides a means of generating multiple transformed images from the same source or applying multiple transformations simultaneously, in order to eliminate the need to read the source coefficients multiple times.

@param handle handle to a TurboJPEG instance that has been initialized for lossless transformation

@param jpegBuf pointer to a byte buffer containing the JPEG source image to transform

@param jpegSize size of the JPEG source image (in bytes)

@param n the number of transformed JPEG images to generate

@param dstBufs pointer to an array of n byte buffers. dstBufs[i] will receive a JPEG image that has been transformed using the parameters in transforms[i]. TurboJPEG has the ability to reallocate the JPEG destination buffer to accommodate the size of the transformed JPEG image. Thus, you can choose to: -# pre-allocate the JPEG destination buffer with an arbitrary size using #tj3Alloc() and let TurboJPEG grow the buffer as needed, -# set dstBufs[i] to NULL to tell TurboJPEG to allocate the buffer for you, or -# pre-allocate the buffer to a “worst case” size determined by calling #tj3TransformBufSize(). Under normal circumstances, this should ensure that the buffer never has to be re-allocated. (Setting #TJPARAM_NOREALLOC guarantees that it won’t be.) Note, however, that there are some rare cases (such as transforming images with a large amount of embedded Exif data) in which the transformed JPEG image will be larger than the worst-case size, and #TJPARAM_NOREALLOC cannot be used in those cases unless the embedded data is discarded using #TJXOPT_COPYNONE or #TJPARAM_SAVEMARKERS. . If you choose option 1 or 3, then dstSizes[i] should be set to the size of your pre-allocated buffer. In any case, unless you have set #TJPARAM_NOREALLOC, you should always check dstBufs[i] upon return from this function, as it may have changed.

@param dstSizes pointer to an array of n size_t variables that will receive the actual sizes (in bytes) of each transformed JPEG image. If dstBufs[i] points to a pre-allocated buffer, then dstSizes[i] should be set to the size of the buffer. Upon return, dstSizes[i] will contain the size of the transformed JPEG image (in bytes.)

@param transforms pointer to an array of n #tjtransform structures, each of which specifies the transform parameters and/or cropping region for the corresponding transformed JPEG image.

@return 0 if successful, or -1 if an error occurred (see #tj3GetErrorStr() and #tj3GetErrorCode().)