turbojpeg_sys

Function tj3TransformBufSize

pub unsafe extern "C" fn tj3TransformBufSize(
    handle: tjhandle,
    transform: *const tjtransform,
) -> size_t

The maximum size of the buffer (in bytes) required to hold a JPEG image transformed with the given transform parameters and/or cropping region. This function is a wrapper for #tj3JPEGBufSize() that takes into account cropping, transposition of the width and height (which affects the destination image dimensions and level of chrominance subsampling), grayscale conversion, and the ICC profile (if any) that was previously associated with the TurboJPEG instance (see #tj3SetICCProfile()) or extracted from the source image (see #tj3GetICCProfile() and #TJPARAM_SAVEMARKERS.) The JPEG header must be read (see tj3DecompressHeader()) prior to calling this function.

@param handle handle to a TurboJPEG instance that has been initialized for lossless transformation

@param transform pointer to a #tjtransform structure that specifies the transform parameters and/or cropping region for the JPEG image.

@return the maximum size of the buffer (in bytes) required to hold the transformed image, or 0 if an error occurred (see #tj3GetErrorStr() and #tj3GetErrorCode().)