turbojpeg_sys

Function tj3YUVBufSize

pub unsafe extern "C" fn tj3YUVBufSize(
    width: c_int,
    align: c_int,
    height: c_int,
    subsamp: c_int,
) -> size_t

The size of the buffer (in bytes) required to hold a unified planar YUV image with the given parameters.

@param width width (in pixels) of the image

@param align row alignment (in bytes) of the image (must be a power of 2.) Setting this parameter to n specifies that each row in each plane of the image will be padded to the nearest multiple of n bytes (1 = unpadded.)

@param height height (in pixels) of the image

@param subsamp level of chrominance subsampling in the image (see @ref TJSAMP “Chrominance subsampling options”.)

@return the size of the buffer (in bytes) required to hold the image, or 0 if the arguments are out of bounds.