turbojpeg_sys

Function tj3YUVPlaneHeight

pub unsafe extern "C" fn tj3YUVPlaneHeight(
    componentID: c_int,
    height: c_int,
    subsamp: c_int,
) -> c_int

The plane height of a YUV image plane with the given parameters. Refer to @ref YUVnotes “YUV Image Format Notes” for a description of plane height.

@param componentID ID number of the image plane (0 = Y, 1 = U/Cb, 2 = V/Cr)

@param height height (in pixels) of the YUV image

@param subsamp level of chrominance subsampling in the image (see @ref TJSAMP “Chrominance subsampling options”.)

@return the plane height of a YUV image plane with the given parameters, or 0 if the arguments are out of bounds.